+++ pytket/binders/circuit/main.cpp
@@ -25,6 +25,7 @@
 #include "tket/Circuit/Command.hpp"
 #include "tket/Gate/OpPtrFunctions.hpp"
 #include "tket/Gate/SymTable.hpp"
+#include "tket/OpType/OpType.hpp"
 #include "tket/Ops/BarrierOp.hpp"
 #include "tket/Ops/MetaOp.hpp"
 #include "tket/Ops/Op.hpp"
@@ -42,6 +43,7 @@ typedef py::tket_custom::SequenceVec<UnitID> py_unit_vector_t;
 void def_circuit(py::class_<Circuit, std::shared_ptr<Circuit>> &);
 void init_classical(py::module &m);
 void init_boxes(py::module &m);
+void init_clexpr(py::module &m);
 
 PYBIND11_MODULE(circuit, m) {
   py::module::import("pytket._tket.unit_id");
@@ -545,6 +547,7 @@ PYBIND11_MODULE(circuit, m) {
           "DiagonalBox", OpType::DiagonalBox,
           "A box for synthesising a diagonal unitary matrix into a sequence of "
           "multiplexed-Rz gates")
+      .value("ClExpr", OpType::ClExpr, "A classical expression")
       .def_static(
           "from_name",
           [](const py::str &name) { return json(name).get<OpType>(); },
@@ -688,6 +691,7 @@ PYBIND11_MODULE(circuit, m) {
       "result in bit 0");
   init_boxes(m);
   init_classical(m);
+  init_clexpr(m);
   def_circuit(pyCircuit);
 
   m.def(