+++ pytket/pytket/qasm/qasm.py
@@ -54,6 +54,9 @@ from pytket._tket.circuit import (
     MultiBitOp,
     WASMOp,
     BarrierOp,
+    ClExprOp,
+    WiredClExpr,
+    ClExpr,
 )
 from pytket._tket.unit_id import _TEMP_BIT_NAME, _TEMP_BIT_REG_BASE
 from pytket.circuit import (
@@ -65,6 +68,11 @@ from pytket.circuit import (
     Qubit,
     QubitRegister,
     UnitID,
+)
+from pytket.circuit.clexpr import (
+    check_register_alignments,
+    has_reg_output,
+    wired_clexpr_from_logic_exp,
 )
 from pytket.circuit.decompose_classical import int_to_bools
 from pytket.circuit.logic_exp import (
@@ -298,7 +306,12 @@ class ParsMap:
 
 
 class CircuitTransformer(Transformer):
-    def __init__(self, return_gate_dict: bool = False, maxwidth: int = 32) -> None:
+    def __init__(
+        self,
+        return_gate_dict: bool = False,
+        maxwidth: int = 32,
+        use_clexpr: bool = False,
+    ) -> None:
         super().__init__()
         self.q_registers: Dict[str, int] = {}
         self.c_registers: Dict[str, int] = {}
@@ -307,6 +320,7 @@ class CircuitTransformer(Transformer):
         self.include = ""
         self.return_gate_dict = return_gate_dict
         self.maxwidth = maxwidth
+        self.use_clexpr = use_clexpr
 
     def _fresh_temp_bit(self) -> List:
         if _TEMP_BIT_NAME in self.c_registers:
@@ -715,6 +729,29 @@ class CircuitTransformer(Transformer):
             },
         }
 
+    def _clexpr_dict(self, exp: LogicExp, out_args: List[List]) -> CommandDict:
+        # Convert the LogicExp to a serialization of a command containing the
+        # corresponding ClExprOp.
+        wexpr, args = wired_clexpr_from_logic_exp(
+            exp, [Bit.from_list(arg) for arg in out_args]
+        )
+        return {
+            "op": {
+                "type": "ClExpr",
+                "expr": wexpr.to_dict(),
+            },
+            "args": [arg.to_list() for arg in args],
+        }
+
+    def _logic_exp_as_cmd_dict(
+        self, exp: LogicExp, out_args: List[List]
+    ) -> CommandDict:
+        return (
+            self._clexpr_dict(exp, out_args)
+            if self.use_clexpr
+            else self._cexpbox_dict(exp, out_args)
+        )
+
     def assign(self, tree: List) -> Iterable[CommandDict]:
         child_iter = iter(tree)
         out_args = list(next(child_iter))
@@ -752,7 +789,7 @@ class CircuitTransformer(Transformer):
         args = args_uids[0]
         if isinstance(out_arg, List):
             if isinstance(exp, LogicExp):
-                yield self._cexpbox_dict(exp, args)
+                yield self._logic_exp_as_cmd_dict(exp, args)
             elif isinstance(exp, (int, bool)):
                 assert exp in (0, 1, True, False)
                 yield {
@@ -769,9 +806,9 @@ class CircuitTransformer(Transformer):
         else:
             reg = out_arg
             if isinstance(exp, RegLogicExp):
-                yield self._cexpbox_dict(exp, args)
+                yield self._logic_exp_as_cmd_dict(exp, args)
             elif isinstance(exp, BitLogicExp):
-                yield self._cexpbox_dict(exp, args[:1])
+                yield self._logic_exp_as_cmd_dict(exp, args[:1])
             elif isinstance(exp, int):
                 yield {
                     "args": args,
@@ -926,38 +963,42 @@ class CircuitTransformer(Transformer):
         return outdict
 
 
-def parser(maxwidth: int) -> Lark:
+def parser(maxwidth: int, use_clexpr: bool) -> Lark:
     return Lark(
         grammar,
         start="prog",
         debug=False,
         parser="lalr",
         cache=True,
-        transformer=CircuitTransformer(maxwidth=maxwidth),
+        transformer=CircuitTransformer(maxwidth=maxwidth, use_clexpr=use_clexpr),
     )
 
 
 g_parser = None
 g_maxwidth = 32
+g_use_clexpr = False
 
 
-def set_parser(maxwidth: int) -> None:
-    global g_parser, g_maxwidth
-    if (g_parser is None) or (g_maxwidth != maxwidth):  # type: ignore
-        g_parser = parser(maxwidth=maxwidth)
+def set_parser(maxwidth: int, use_clexpr: bool) -> None:
+    global g_parser, g_maxwidth, g_use_clexpr
+    if (g_parser is None) or (g_maxwidth != maxwidth) or g_use_clexpr != use_clexpr:  # type: ignore
+        g_parser = parser(maxwidth=maxwidth, use_clexpr=use_clexpr)
         g_maxwidth = maxwidth
+        g_use_clexpr = use_clexpr
 
 
 def circuit_from_qasm(
     input_file: Union[str, "os.PathLike[Any]"],
     encoding: str = "utf-8",
     maxwidth: int = 32,
+    use_clexpr: bool = False,
 ) -> Circuit:
     """A method to generate a tket Circuit from a qasm file.
 
     :param input_file: path to qasm file; filename must have ``.qasm`` extension
     :param encoding: file encoding (default utf-8)
     :param maxwidth: maximum allowed width of classical registers (default 32)
+    :param use_clexpr: whether to use ClExprOp to represent classical expressions
     :return: pytket circuit
     """
     ext = os.path.splitext(input_file)[-1]
@@ -965,21 +1006,24 @@ def circuit_from_qasm(
         raise TypeError("Can only convert .qasm files")
     with open(input_file, "r", encoding=encoding) as f:
         try:
-            circ = circuit_from_qasm_io(f, maxwidth=maxwidth)
+            circ = circuit_from_qasm_io(f, maxwidth=maxwidth, use_clexpr=use_clexpr)
         except QASMParseError as e:
             raise QASMParseError(e.msg, e.line, str(input_file))
     return circ
 
 
-def circuit_from_qasm_str(qasm_str: str, maxwidth: int = 32) -> Circuit:
+def circuit_from_qasm_str(
+    qasm_str: str, maxwidth: int = 32, use_clexpr: bool = False
+) -> Circuit:
     """A method to generate a tket Circuit from a qasm string.
 
     :param qasm_str: qasm string
     :param maxwidth: maximum allowed width of classical registers (default 32)
+    :param use_clexpr: whether to use ClExprOp to represent classical expressions
     :return: pytket circuit
     """
     global g_parser
-    set_parser(maxwidth=maxwidth)
+    set_parser(maxwidth=maxwidth, use_clexpr=use_clexpr)
     assert g_parser is not None
     cast(CircuitTransformer, g_parser.options.transformer)._reset_context(
         reset_wasm=False
@@ -987,9 +1031,13 @@ def circuit_from_qasm_str(qasm_str: str, maxwidth: int = 32) -> Circuit:
     return Circuit.from_dict(g_parser.parse(qasm_str))  # type: ignore[arg-type]
 
 
-def circuit_from_qasm_io(stream_in: TextIO, maxwidth: int = 32) -> Circuit:
+def circuit_from_qasm_io(
+    stream_in: TextIO, maxwidth: int = 32, use_clexpr: bool = False
+) -> Circuit:
     """A method to generate a tket Circuit from a qasm text stream"""
-    return circuit_from_qasm_str(stream_in.read(), maxwidth=maxwidth)
+    return circuit_from_qasm_str(
+        stream_in.read(), maxwidth=maxwidth, use_clexpr=use_clexpr
+    )
 
 
 def circuit_from_qasm_wasm(
@@ -997,6 +1045,7 @@ def circuit_from_qasm_wasm(
     wasm_file: Union[str, "os.PathLike[Any]"],
     encoding: str = "utf-8",
     maxwidth: int = 32,
+    use_clexpr: bool = False,
 ) -> Circuit:
     """A method to generate a tket Circuit from a qasm string and external WASM module.
 
@@ -1008,10 +1057,12 @@ def circuit_from_qasm_wasm(
     """
     global g_parser
     wasm_module = WasmFileHandler(str(wasm_file))
-    set_parser(maxwidth=maxwidth)
+    set_parser(maxwidth=maxwidth, use_clexpr=use_clexpr)
     assert g_parser is not None
     cast(CircuitTransformer, g_parser.options.transformer).wasm = wasm_module
-    return circuit_from_qasm(input_file, encoding=encoding, maxwidth=maxwidth)
+    return circuit_from_qasm(
+        input_file, encoding=encoding, maxwidth=maxwidth, use_clexpr=use_clexpr
+    )
 
 
 def circuit_to_qasm(
@@ -1082,13 +1133,16 @@ def check_can_convert_circuit(circ: Circuit, header: str, maxwidth: int) -> None
             f"Circuit contains a classical register larger than {maxwidth}: try "
             "setting the `maxwidth` parameter to a higher value."
         )
+    # Empty CustomGates should have been removed by DecomposeBoxes().
     for cmd in circ:
-        if is_empty_customgate(cmd.op) or (
-            isinstance(cmd.op, Conditional) and is_empty_customgate(cmd.op.op)
-        ):
-            raise QASMUnsupportedError(
-                f"Empty CustomGates and opaque gates are not supported."
-            )
+        assert not is_empty_customgate(cmd.op)
+        if isinstance(cmd.op, Conditional):
+            assert not is_empty_customgate(cmd.op.op)
+    if not check_register_alignments(circ):
+        raise QASMUnsupportedError(
+            "Circuit contains classical expressions on registers whose arguments or "
+            "outputs are not register-aligned."
+        )
 
 
 def circuit_to_qasm_str(
@@ -1111,12 +1165,12 @@ def circuit_to_qasm_str(
     :return: qasm string
     """
 
-    check_can_convert_circuit(circ, header, maxwidth)
     qasm_writer = QasmWriter(
         circ.qubits, circ.bits, header, include_gate_defs, maxwidth
     )
     circ1 = circ.copy()
     DecomposeBoxes().apply(circ1)
+    check_can_convert_circuit(circ1, header, maxwidth)
     for command in circ1:
         assert isinstance(command, Command)
         qasm_writer.add_op(command.op, command.args)
@@ -1471,25 +1525,27 @@ class QasmWriter:
         else:
             self.variable_writes[label] = [written_variable]
 
-    def add_range_predicate(self, op: RangePredicateOp, args: List[Bit]) -> None:
-        comparator, value = _parse_range(op.lower, op.upper, self.maxwidth)
-        if (not hqs_header(self.header)) and comparator != "==":
+    def check_range_predicate(self, op: RangePredicateOp, args: List[Bit]) -> None:
+        if (not hqs_header(self.header)) and op.lower != op.upper:
             raise QASMUnsupportedError(
                 "OpenQASM conditions must be on a register's fixed value."
             )
-        bits = args[:-1]
+        variable = args[0].reg_name
+        assert isinstance(variable, str)
+        if op.n_inputs != self.cregs[variable].size:
+            raise QASMUnsupportedError(
+                "RangePredicate conditions must be an entire classical register"
+            )
+        if args[:-1] != self.cregs[variable].to_list():
+            raise QASMUnsupportedError(
+                "RangePredicate conditions must be a single classical register"
+            )
+
+    def add_range_predicate(self, op: RangePredicateOp, args: List[Bit]) -> None:
+        self.check_range_predicate(op, args)
+        comparator, value = _parse_range(op.lower, op.upper, self.maxwidth)
         variable = args[0].reg_name
         dest_bit = str(args[-1])
-        if not hqs_header(self.header):
-            assert isinstance(variable, str)
-            if op.n_inputs != self.cregs[variable].size:
-                raise QASMUnsupportedError(
-                    "OpenQASM conditions must be an entire classical register"
-                )
-            if bits != self.cregs[variable].to_list():
-                raise QASMUnsupportedError(
-                    "OpenQASM conditions must be a single classical register"
-                )
         label = self.strings.add_string(
             "".join(
                 [
@@ -1606,9 +1662,42 @@ class QasmWriter:
             # Conditional phase is ignored.
             return
         if op.op.type == OpType.RangePredicate:
-            raise QASMUnsupportedError(
-                "Conditional RangePredicate is currently unsupported."
+            # Special handling for nested ifs
+            # if condition
+            #   if pred dest = 1
+            #   if not pred dest = 0
+            # can be written as
+            # if condition s0 = 1
+            # if pred s1 = 1
+            # s2 = s0 & s1
+            # s3 = s0 & ~s1
+            # if s2 dest = 1
+            # if s3 dest = 0
+            # where s0, s1, s2, and s3 are scratch bits
+            s0 = self.fresh_scratch_bit()
+            l = self.strings.add_string(f"{s0} = 1;\n")
+            # we store the condition in self.strings.conditions
+            # as it can be later replaced by `replace_condition`
+            # if possible
+            self.strings.conditions[l] = ConditionString(variable, "==", op.value)
+            # output the RangePredicate to s1
+            s1 = self.fresh_scratch_bit()
+            assert isinstance(op.op, RangePredicateOp)
+            self.check_range_predicate(op.op, cast(List[Bit], args[op.width :]))
+            pred_comparator, pred_value = _parse_range(
+                op.op.lower, op.op.upper, self.maxwidth
             )
+            pred_variable = args[op.width :][0].reg_name
+            self.strings.add_string(
+                f"if({pred_variable}{pred_comparator}{pred_value}) {s1} = 1;\n"
+            )
+            s2 = self.fresh_scratch_bit()
+            self.strings.add_string(f"{s2} = {s0} & {s1};\n")
+            s3 = self.fresh_scratch_bit()
+            self.strings.add_string(f"{s3} = {s0} & (~ {s1});\n")
+            self.strings.add_string(f"if({s2}==1) {args[-1]} = 1;\n")
+            self.strings.add_string(f"if({s3}==1) {args[-1]} = 0;\n")
+            return
         # we assign the condition to a scratch bit, which we will later remove
         # if the condition variable is unchanged.
         scratch_bit = self.fresh_scratch_bit()
@@ -1718,6 +1807,53 @@ class QasmWriter:
                 " for writing to a single bit or whole registers."
             )
 
+    def add_wired_clexpr(self, op: ClExprOp, args: List[Bit]) -> None:
+        wexpr: WiredClExpr = op.expr
+        # 1. Determine the mappings from bit variables to bits and from register
+        # variables to registers.
+        expr: ClExpr = wexpr.expr
+        bit_posn: dict[int, int] = wexpr.bit_posn
+        reg_posn: dict[int, list[int]] = wexpr.reg_posn
+        output_posn: list[int] = wexpr.output_posn
+        input_bits: dict[int, Bit] = {i: args[j] for i, j in bit_posn.items()}
+        input_regs: dict[int, BitRegister] = {}
+        all_cregs = set(self.cregs.values())
+        for i, posns in reg_posn.items():
+            reg_args = [args[j] for j in posns]
+            for creg in all_cregs:
+                if creg.to_list() == reg_args:
+                    input_regs[i] = creg
+                    break
+            else:
+                assert (
+                    not f"ClExprOp ({wexpr}) contains a register variable (r{i}) that "
+                    "is not wired to any BitRegister in the circuit."
+                )
+        # 2. Write the left-hand side of the assignment.
+        output_repr: Optional[str] = None
+        output_args: list[Bit] = [args[j] for j in output_posn]
+        n_output_args = len(output_args)
+        expect_reg_output = has_reg_output(expr.op)
+        if n_output_args == 0:
+            raise QASMUnsupportedError("Expression has no output.")
+        elif n_output_args == 1:
+            output_arg = output_args[0]
+            output_repr = output_arg.reg_name if expect_reg_output else str(output_arg)
+        else:
+            if not expect_reg_output:
+                raise QASMUnsupportedError("Unexpected output for operation.")
+            for creg in all_cregs:
+                if creg.to_list() == output_args:
+                    output_repr = creg.name
+                    break
+            assert output_repr is not None
+        self.strings.add_string(f"{output_repr} = ")
+        # 3. Write the right-hand side of the assignment.
+        self.strings.add_string(
+            expr.as_qasm(input_bits=input_bits, input_regs=input_regs)
+        )
+        self.strings.add_string(";\n")
+
     def add_wasm(self, op: WASMOp, args: List[Bit]) -> None:
         inputs: List[str] = []
         outputs: List[str] = []
@@ -1821,6 +1957,9 @@ class QasmWriter:
         elif optype == OpType.ClassicalExpBox:
             assert isinstance(op, ClassicalExpBox)
             self.add_classical_exp_box(op, cast(List[Bit], args))
+        elif optype == OpType.ClExpr:
+            assert isinstance(op, ClExprOp)
+            self.add_wired_clexpr(op, cast(List[Bit], args))
         elif optype == OpType.WASM:
             assert isinstance(op, WASMOp)
             self.add_wasm(op, cast(List[Bit], args))