+++ pytket/tests/qasm_test.py
@@ -422,7 +422,7 @@ def test_h1_rzz() -> None:
 def test_extended_qasm() -> None:
     fname = str(curr_file_path / "qasm_test_files/test17.qasm")
     out_fname = str(curr_file_path / "qasm_test_files/test17_output.qasm")
-    c = circuit_from_qasm_wasm(fname, "testfile.wasm")
+    c = circuit_from_qasm_wasm(fname, "testfile.wasm", use_clexpr=True)
 
     out_qasm = circuit_to_qasm_str(c, "hqslib1")
     with open(out_fname) as f:
@@ -432,15 +432,14 @@ def test_extended_qasm() -> None:
 
     assert circuit_to_qasm_str(c2, "hqslib1")
 
-    with pytest.raises(DecomposeClassicalError) as e:
-        DecomposeClassicalExp().apply(c)
+    assert not DecomposeClassicalExp().apply(c)
 
 
 def test_decomposable_extended() -> None:
     fname = str(curr_file_path / "qasm_test_files/test18.qasm")
     out_fname = str(curr_file_path / "qasm_test_files/test18_output.qasm")
 
-    c = circuit_from_qasm_wasm(fname, "testfile.wasm", maxwidth=64)
+    c = circuit_from_qasm_wasm(fname, "testfile.wasm", maxwidth=64, use_clexpr=True)
     DecomposeClassicalExp().apply(c)
 
     out_qasm = circuit_to_qasm_str(c, "hqslib1", maxwidth=64)
@@ -452,9 +451,9 @@ def test_opaque() -> None:
     c = circuit_from_qasm_str(
         'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[4];\nopaque myopaq() q1, q2;\n myopaq() q[0], q[1];'
     )
-    with pytest.raises(QASMUnsupportedError) as e:
-        circuit_to_qasm_str(c)
-    assert "Empty CustomGates and opaque gates are not supported" in str(e.value)
+    assert (
+        circuit_to_qasm_str(c) == 'OPENQASM 2.0;\ninclude "qelib1.inc";\n\nqreg q[4];\n'
+    )
 
 
 def test_alternate_encoding() -> None:
@@ -654,10 +653,11 @@ def test_qasm_phase() -> None:
         assert c1 == c0
 
 
-def test_CopyBits() -> None:
+@pytest.mark.parametrize("use_clexpr", [True, False])
+def test_CopyBits(use_clexpr: bool) -> None:
     input_qasm = """OPENQASM 2.0;\ninclude "hqslib1.inc";\n\ncreg c0[1];
 creg c1[3];\nc0[0] = c1[1];\n"""
-    c = circuit_from_qasm_str(input_qasm)
+    c = circuit_from_qasm_str(input_qasm, use_clexpr=use_clexpr)
     result_circ_qasm = circuit_to_qasm_str(c, "hqslib1")
     assert input_qasm == result_circ_qasm
 
@@ -831,7 +831,8 @@ measure q[0] -> c[32];"""
     assert len(circ_out.bits) == 33
 
 
-def test_classical_expbox_arg_order() -> None:
+@pytest.mark.parametrize("use_clexpr", [True, False])
+def test_classical_expbox_arg_order(use_clexpr: bool) -> None:
     qasm = """
     OPENQASM 2.0;
     include "hqslib1.inc";
@@ -846,7 +847,7 @@ def test_classical_expbox_arg_order() -> None:
     c = a ^ b | d;
     """
 
-    circ = circuit_from_qasm_str(qasm)
+    circ = circuit_from_qasm_str(qasm, use_clexpr=use_clexpr)
     args = circ.get_commands()[0].args
     expected_symbol_order = ["a", "b", "d", "c"]
     expected_index_order = [0, 1, 2, 3]
@@ -1025,16 +1026,62 @@ if(tk_SCRATCH_BITREG_0[0]==1) c[1] = 1;
 
 
 def test_conditional_range_predicate() -> None:
-    range_predicate = RangePredicateOp(6, 0, 27)
-    c = Circuit(0, 8)
-    c.add_gate(range_predicate, [0, 1, 2, 3, 4, 5, 6], condition=Bit(7))
-    # remove once https://github.com/CQCL/tket/issues/1508
-    # is resolved
+    range_predicate = RangePredicateOp(2, 0, 2)
+    c = Circuit(0, 5)
+    c.add_gate(range_predicate, [1, 2, 4])
+    # https://github.com/CQCL/tket/issues/1642
     with pytest.raises(Exception) as errorinfo:
-        circuit_to_qasm_str(c, header="hqslib1")
-        assert "Conditional RangePredicate is currently unsupported." in str(
+        qasm = circuit_to_qasm_str(c, header="hqslib1")
+        assert "RangePredicate conditions must be an entire classical register" in str(
             errorinfo.value
         )
+    # https://github.com/CQCL/tket/issues/1508
+    range_predicate = RangePredicateOp(6, 0, 27)
+    c = Circuit(0, 6)
+    c.add_gate(range_predicate, [0, 1, 2, 3, 4, 5, 5], condition=Bit(5))
+    qasm = circuit_to_qasm_str(c, header="hqslib1")
+    assert (
+        qasm
+        == """OPENQASM 2.0;
+include "hqslib1.inc";
+
+creg c[6];
+creg tk_SCRATCH_BITREG_0[4];
+if(c[5]==1) tk_SCRATCH_BITREG_0[0] = 1;
+if(c<=27) tk_SCRATCH_BITREG_0[1] = 1;
+tk_SCRATCH_BITREG_0[2] = tk_SCRATCH_BITREG_0[0] & tk_SCRATCH_BITREG_0[1];
+tk_SCRATCH_BITREG_0[3] = tk_SCRATCH_BITREG_0[0] & (~ tk_SCRATCH_BITREG_0[1]);
+if(tk_SCRATCH_BITREG_0[2]==1) c[5] = 1;
+if(tk_SCRATCH_BITREG_0[3]==1) c[5] = 0;
+"""
+    )
+    # more test
+    range_predicate = RangePredicateOp(2, 0, 2)
+    c = Circuit()
+    reg_a = c.add_c_register("a", 2)
+    reg_b = c.add_c_register("b", 2)
+    reg_d = c.add_c_register("d", 1)
+    c.add_gate(
+        range_predicate, reg_a.to_list() + reg_d.to_list(), condition=reg_gt(reg_b, 1)
+    )
+    qasm = circuit_to_qasm_str(c, header="hqslib1")
+    assert (
+        qasm
+        == """OPENQASM 2.0;
+include "hqslib1.inc";
+
+creg a[2];
+creg b[2];
+creg d[1];
+creg tk_SCRATCH_BITREG_0[4];
+if(b>=2) tk_SCRATCH_BITREG_0[0] = 1;
+if(a<=2) tk_SCRATCH_BITREG_0[1] = 1;
+tk_SCRATCH_BITREG_0[2] = tk_SCRATCH_BITREG_0[0] & tk_SCRATCH_BITREG_0[1];
+tk_SCRATCH_BITREG_0[3] = tk_SCRATCH_BITREG_0[0] & (~ tk_SCRATCH_BITREG_0[1]);
+if(tk_SCRATCH_BITREG_0[2]==1) d[0] = 1;
+if(tk_SCRATCH_BITREG_0[3]==1) d[0] = 0;
+"""
+    )
 
 
 def test_range_with_maxwidth() -> None:
@@ -1158,5 +1205,6 @@ if __name__ == "__main__":
     test_header_stops_gate_definition()
     test_tk2_definition()
     test_rxxyyzz_conversion()
-    test_classical_expbox_arg_order()
+    test_classical_expbox_arg_order(True)
+    test_classical_expbox_arg_order(False)
     test_register_name_check()