+++ pytket/docs/changelog.rst
@@ -1,6 +1,45 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Features:
+
+* Add `clexpr.check_register_alignments()` method to check register alignments
+  in `ClExprOp`.
+
+Fixes:
+
+* Fix `symbol_substitution` not preserving opgroups.
+* Remove hardware inefficient circuit construction in `_tk1_to_rzsx`
+* Support converting conditional `RangePredicate`s to QASM.
+
+1.34.0 (October 2024)
+---------------------
+
+Features:
+
+* Add new `ClExprOp` operation type as an alternative to `ClassicalExpBox`; add
+  option to use this when converting from QASM.
+* Several updates to `GreedyPauliSimp`:
+
+  * Support for mid-circuit measurements, resets, conditionals, and classical gates.
+  * New parameters `max_lookahead` and `max_tqe_candidates` are added
+    to limit the search space.
+  * New parameter `seed` is added to support random sampling and tie breaking.
+  * New parameter `allow_zzphase` allows the algorithm to implement 2-qubit rotations
+    using ZZPhase gates when deemed optimal.
+
+Fixes:
+
+* Fix small default display screen for circuit renderer.
+
+General:
+
+* Support Python 3.13.
+* Drop support for MacOS 12.
+
 1.33.1 (October 2024)
 ---------------------